Archade: one listing with volume, users and the full fee picture
Merges the three-product split back into the single archade entry. Volume is
every swap Archade is part of, counted once by signature: curve swaps from any
interface plus app swaps routed to other venues. Fees carry all seven legs;
revenue is the app fee, the curve partner share and the launch fee share;
users and transactions are the union of both ledgers.
